7+ Taskbar Tweaker v5.14.1

+++ dll/prevent_explorer_appid_changes.c
@@ -5,7 +5,7 @@
 
 static MODULEINFO ExplorerFrameInfo;
 static MODULEINFO Shell32Info;
-static volatile int nHookCallCounter;
+static volatile LONG nHookCallCounter;
 
 static void **ppPropertyStoreSetValue;
 POINTER_REDIRECTION_VAR(static POINTER_REDIRECTION prPropertyStoreSetValue);
